fix(docs): fix docs sync workflow
Fixes an issue where the `bun install` command in the docs sync workflow was failing because of the `--frozen-lockfile` flag. This flag requires a `bun.lockb` file to exist, which was not being generated or committed. This change removes the `--frozen-lockfile` flag to allow the workflow to install dependencies correctly. Also added paths to trigger sync on markdown changes.