feat(api): Implement docker status endpoint
This commit introduces a new endpoint `/docker/status` that retrieves and returns the status of the Docker pool. It also adds an `initialized` property to the `WorkerMetrics` interface and sets it in the `DockerClientManager`. The docker client init function is now called upon initialisation. This helps ensure that the API can accurately report the health and state of the Docker environment.

The commit includes:
- Implementation of the `/docker/status` endpoint in the `dockstat` app.
- Addition of the `initialized` property to the `WorkerMetrics` interface in the `docker-client` package.
- Setting `initialized` value.
- Calling init function.