feat(core): Add plugin events and worker event proxy
This commit introduces a new feature that allows plugins to register event handlers for Docker client events. This is achieved by:

- Adding a new `events` property to the `Plugin` interface in `@dockstat/typings`.
- Updating the `DockerClientManager` in `@dockstat/docker-client` to listen for events from workers and trigger the corresponding plugin event handlers.
- Adding a new `proxyEvent` function in `@dockstat/docker-client` to proxy events from the worker to the main process.
- Adding a new `@dockstat/utils` package to provide utility functions for building messages.
- Upgrading bun-types to 1.3.3

The new `events` property in the `Plugin` interface allows plugins to specify a set of event handlers that will be called when the corresponding Docker client events are triggered.

The `DockerClientManager` has been updated to listen for events from workers. When an event is received, the `DockerClientManager` will iterate over the registered plugins and call the corresponding event handler, if one exists.

The `proxyEvent` function is used to proxy events from the worker to the main process. This function takes the event type, the event context, and an optional additional context as arguments. It then posts a message to the main process containing the event data.

The `@dockstat/utils` package provides utility functions for building messages. This package is used by the `proxyEvent` function to build the message that is posted to the main process.